gem join

+++ src/interfaces/ICdpManager.sol
@@ -3,7 +3,7 @@ pragma solidity 0.8.19;
 
 interface ICdpManager {
     function vat() external view returns (address);
-    function safes(uint) external view returns (address);
-    function collateralTypes(uint) external view returns (bytes32);
+    function safes(uint256) external view returns (address);
+    function collateralTypes(uint256) external view returns (bytes32);
     function open(bytes32, address) external returns (uint256);
 }
